Update Avrorouter tutorial
The tutorial now displays a single configuration example which should be easier
to understand.

Added a sub-section about configuring the master server with proper replication
settings and explained the need for the CREATE TABLE statements in the binary
logs.

Display an example Avro JSON schema and provide links to the schema definition
documents. Mention the cdc_schema utility as the second option for creating the
schema files.